feat: add 404 and 503 error pages; refactor error handling components and remove obsolete styles
