+++ client/src/App.jsx
@@ -8,14 +8,19 @@ import ErrorBoundary from './middleware/ErrorBoundary';
 import { LoadingProvider } from './contexts/LoadingContext';
 
 
+// App component
 export default function App() {
 
   return (
     <ErrorBoundary fallback={<ErrorPage />}>
+
       <LoadingProvider>
-        <Loading />
-        <Router />
+        <Loading /> {/* Loading component */}
+
+        <Router /> {/* Routing Pages */}
       </LoadingProvider>
+
+      {/* Toastify */}
       <ToastContainer
         theme="colored"
         draggable={false}
@@ -23,6 +28,7 @@ export default function App() {
         hideProgressBar={true}
         position="bottom-right"
       />
+
     </ErrorBoundary>
   );
 };
