+++ client/src/pages/503/ServerUnavl.jsx
@@ -4,26 +4,27 @@ import {
 } from "@mui/material";
 import { DotLottieReact } from "@lottiefiles/dotlottie-react";
 
-export default function PageNotFound() {
+
+// Server Unavailable component
+export default function ServerUnavl() {
     return (
         <Container maxWidth="lg" sx={{ height: '100vh', textAlign: 'center' }}>
             <Typography variant="h3" pt={8}>
-                404 - Page Not Found
+                503 - Service Unavailable
             </Typography>
             <Typography variant="h5" py={1}>
-                The page you are looking for does not exist.
+                The server is currently unavailable. Please try again later.
             </Typography>
             <Button variant="contained" component={Link} to="/" sx={{ marginY: "30px" }}>
                 Go to Home
             </Button>
             <div style={{ height: '50vh', width: '100%' }}>
                 <DotLottieReact
-                    src="PageNotFound.json"
+                    src="ServerUnavl.json"
                     loop={true}
                     autoplay={true}
                     style={{ width: '100%', height: '100%' }}
                 />
-
             </div>
         </Container>
     );
