--- client/src/pages/error/ErrorPage.css
@@ -1,34 +1,0 @@
-.error-page {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: 100vh;
-    text-align: center;
-    background-color: #f8f9fa;
-}
-
-.error-page h1 {
-    font-size: 2.5rem;
-    color: #dc3545;
-}
-
-.error-page p {
-    margin: 1rem 0;
-    font-size: 1.2rem;
-    color: #6c757d;
-}
-
-.error-page button {
-    padding: 0.5rem 1rem;
-    font-size: 1rem;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-}
-
-.error-page button:hover {
-    background-color: #0056b3;
-}
